w3c-reports/12092014/mac/runned-tests/navigation-timing/test_readwrite-expected.txt

Description

This test validates that the window.performance object is read/write.


FAIL window.performance is defined and not null. assert_true: window.performance is defined and not null expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/navigation-timing/resources/webperftestharness.js:58:42
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
wp_test@http://localhost:8800/navigation-timing/resources/webperftestharness.js:58:17
test_namespace@http://localhost:8800/navigation-timing/resources/webperftestharness.js:69:16
global code@http://localhost:8800/navigation-timing/test_readwrite.html:19:27)
FAIL window.performance is defined assert_true: window.performance is defined expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/navigation-timing/resources/webperftestharness.js:69:42
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
wp_test@http://localhost:8800/navigation-timing/resources/webperftestharness.js:62:9
test_namespace@http://localhost:8800/navigation-timing/resources/webperftestharness.js:69:16
global code@http://localhost:8800/navigation-timing/test_readwrite.html:19:27)
PASS window.performance is read/write 
PASS var performance is read/write 
